move credit command function
